カテゴリーページ作成

+++ src/components/organisms/partner/ShowPartnerList.tsx
@@ -4,7 +4,6 @@ import ContentsForm from '@/components/organisms/common/ContentsForm'
 import ListArea from '@/components/organisms/common/ListArea'
 import PartnerList from '@/components/molcules/partner/PartnerList'
 import FormTitle from '@/components/atoms/FormTitle'
-import Balloon from '@/components/atoms/Balloon'
 
 const ShowPartnerList: FC = () => {
   const { partners, deletePartner } = usePartner()
@@ -14,23 +13,27 @@ const ShowPartnerList: FC = () => {
    * @param id
    */
   const handleSubmitDelete = async (id: number) => {
-    await deletePartner(id)
+    if (confirm('パートナーを削除しますか？')) {
+      await deletePartner(id)
+    }
   }
 
   return (
     <ContentsForm>
       <FormTitle title="共有パートナー一覧" space="sm" />
-      <ListArea>
-        {partners.map((partner) => (
-          <PartnerList
-            key={partner.id}
-            id={partner.id}
-            name={partner.name}
-            email={partner.email}
-            submit={handleSubmitDelete}
-          />
-        ))}
-      </ListArea>
+      {partners.length !== 0 && partners[0].id !== 0 && (
+        <ListArea>
+          {partners.map((partner) => (
+            <PartnerList
+              key={partner.id}
+              id={partner.id}
+              name={partner.name}
+              email={partner.email}
+              submit={handleSubmitDelete}
+            />
+          ))}
+        </ListArea>
+      )}
     </ContentsForm>
   )
 }
